requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
numpy==1.26.4
pymongo==4.6.1
python-dotenv==1.0.1
Pillow==10.2.0
gtts==2.3.2
SpeechRecognition==3.10.0
pygame==2.5.0
protobuf==3.20.3
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
pyaudio==0.2.13
openai-whisper==20240930
pydub==0.25.1
watchdog==3.0.0
tensorflow==2.15.0
torch==2.2.0
torchaudio==2.2.0
reportlab==4.1.0
pandas==2.2.0
scikit-learn==1.4.0
matplotlib==3.8.2
joblib==1.3.2
